Changed message type to signed 16 bit int. Might be some unkown changes made yesterday, hopefully not

+++ assign1_ws/install/assign1_interfaces/include/assign1_interfaces/assign1_interfaces/msg/detail/pixel_coordinates__struct.h
@@ -22,8 +22,8 @@ extern "C"
 /// Struct defined in msg/PixelCoordinates in the package assign1_interfaces.
 typedef struct assign1_interfaces__msg__PixelCoordinates
 {
-  uint8_t x;
-  uint8_t y;
+  int16_t x;
+  int16_t y;
 } assign1_interfaces__msg__PixelCoordinates;
 
 // Struct for a sequence of assign1_interfaces__msg__PixelCoordinates.
